--- src/main/java/duke/ExitCommand.java
@@ -1,9 +1,0 @@
-package duke;
-
-public class ExitCommand extends Command {
-    @Override
-    protected void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.printExitMessage();
-        this.setExit();
-    }
-}
