--- src/main/java/duke/HelpCommand.java
@@ -1,8 +1,0 @@
-package duke;
-
-public class HelpCommand extends Command {
-    @Override
-    protected void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.printHelpMessage();
-    }
-}
