--- src/main/java/duke/ListCommand.java
@@ -1,8 +1,0 @@
-package duke;
-
-public class ListCommand extends Command {
-    @Override
-    protected void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.printList(tasks.getAllTasks());
-    }
-}
