+++ src/main/java/duke/Parser.java
@@ -1,5 +1,11 @@
 package duke;
 
+import duke.command.AddCommand;
+import duke.command.Command;
+import duke.command.ExitCommand;
+import duke.command.HelpCommand;
+import duke.command.ListCommand;
+import duke.command.MarkAndDelCommand;
 import duke.exception.EmptyTaskException;
 import duke.exception.IllegalCommandException;
 import duke.exception.InvalidDeadline;
