+++ src/main/java/duke/command/ExitCommand.java
@@ -1,0 +1,13 @@
+package duke.command;
+
+import duke.Storage;
+import duke.task.TaskList;
+import duke.Ui;
+
+public class ExitCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.printExitMessage();
+        this.setExit();
+    }
+}
