+++ src/main/java/duke/command/HelpCommand.java
@@ -1,0 +1,12 @@
+package duke.command;
+
+import duke.Storage;
+import duke.task.TaskList;
+import duke.Ui;
+
+public class HelpCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.printHelpMessage();
+    }
+}
